requirements.txt
flask
opencv-python-headless
face-recognition
firebase-admin
requests
streamlit
numpy
pillow
